♻️ refactor: 알림 내역 관련 코드를 Home에서 History로 이동
