+++ Makcha/src/pages/History/components/CurrentAlarmCard.tsx
