+++ Makcha/src/pages/History/components/EmptyHistoryCard.tsx
