+++ Makcha/src/pages/History/mocks/historyMock.ts
@@ -28,8 +28,8 @@ export type MonthSectionData = {
     items: HistoryItem[];
 };
 
-// export const CURRENT_ALARM_MOCK: CurrentAlarm | null = null;
-export const CURRENT_ALARM_MOCK: CurrentAlarm = {
+export const CURRENT_ALARM_MOCK: CurrentAlarm | null = null;
+/* export const CURRENT_ALARM_MOCK: CurrentAlarm = {
     isOptimal: true,
     lines: ["2호선", "4호선"],
     departureTime: "23:55",
@@ -37,20 +37,20 @@ export const CURRENT_ALARM_MOCK: CurrentAlarm = {
     totalDurationMin: 45,
     transferCount: 1,
     walkingTimeMin: 9,
-};
+}; */
 
 export const PAST_SUMMARY_MOCK: PastSummary = {
-    //thisMonthTaxiCost: 0,
-    //savedCount: 0,
-    thisMonthTaxiCost: 125000,
-    savedCount: 5,
+    thisMonthTaxiCost: 0,
+    savedCount: 0,
+    //thisMonthTaxiCost: 125000,
+    //savedCount: 5,
 };
 
 export const MONTH_SECTIONS_MOCK: MonthSectionData[] = [
     {
         monthLabel: "2025년 12월",
-        //items: [],
-        items: [
+        items: [],
+        /*items: [
             {
                 id: "h1",
                 date: "2025. 12. 03",
@@ -60,6 +60,6 @@ export const MONTH_SECTIONS_MOCK: MonthSectionData[] = [
                 arriveAt: "25:10",
                 savedAmount: 325000,
             },
-        ],
+        ],*/
     },
 ];
